+++ push_swap_files/arg_utils.c
@@ -6,13 +6,13 @@
 /*   By: iyamada <iyamada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/02 01:13:56 by iyamada           #+#    #+#             */
-/*   Updated: 2021/12/30 14:12:10 by iyamada          ###   ########.fr       */
+/*   Updated: 2022/01/04 15:25:47 by iyamada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-static bool	ft_is_exist(int *arry, int size, long long val)
+static bool	ft_is_exist(long long *arry, int size, long long val)
 {
 	int	i;
 
@@ -28,13 +28,13 @@ static bool	ft_is_exist(int *arry, int size, long long val)
 
 static bool	ft_is_wrong_arry(int argc, char *argv[])
 {
-	int		*arry;
-	int		arry_size;
-	int		i;
-	char	*non_num;
+	long long	*arry;
+	int			arry_size;
+	int			i;
+	char		*non_num;
 
 	arry_size = argc - 1;
-	arry = (int *)malloc(arry_size * sizeof(long long));
+	arry = (long long *)malloc(arry_size * sizeof(long long));
 	i = 0;
 	while (i < arry_size)
 	{
